modify usecase
