+++ reader_service/internal/domain/models.go
@@ -1,0 +1,1 @@
+package domain
