+++ writer_service/internal/delivery/grpc/grpc_service_product.go
@@ -6,11 +6,11 @@ import (
 	"github.com/minhwalker/cqrs-microservices/core/pkg/logger"
 	"github.com/minhwalker/cqrs-microservices/core/pkg/tracing"
 	"github.com/minhwalker/cqrs-microservices/writer_service/config"
+	"github.com/minhwalker/cqrs-microservices/writer_service/internal/domain/usecase"
 	"github.com/minhwalker/cqrs-microservices/writer_service/internal/dto"
 	"github.com/minhwalker/cqrs-microservices/writer_service/internal/dto/proto/product_writer"
 	"github.com/minhwalker/cqrs-microservices/writer_service/internal/mappers"
 	product2 "github.com/minhwalker/cqrs-microservices/writer_service/internal/metrics"
-	"github.com/minhwalker/cqrs-microservices/writer_service/internal/usecase/product"
 	uuid "github.com/satori/go.uuid"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
