--- writer_service/internal/usecase/product/get_by_id.go
@@ -1,10 +1,0 @@
-package usecase
-
-import (
-	"context"
-	"github.com/minhwalker/cqrs-microservices/writer_service/internal/domain/models"
-)
-
-func (u *productUsecase) GetByID(ctx context.Context, query *models.GetProductByIdQuery) (*models.Product, error) {
-	return u.productRepository.GetProductById(ctx, query.ProductID)
-}
